* commit '7ea3212da8bf266f99625733297fd52c411e4ce4':
  [EDELIVERY-7880] add javadoc to test class
  [EDELIVERY-7880] remove certificateService reference from crypto service to break cyclic dependency
  [EDELIVERY-7931] fix int overflow when computing the message expiration date

+++ Domibus-MSH/src/main/java/eu/domibus/core/ebms3/sender/retry/UpdateRetryLoggingService.java
@@ -296,7 +296,7 @@ public class UpdateRetryLoggingService {
                                          final LegConfiguration legConfiguration) {
         if (legConfiguration.getReceptionAwareness() != null) {
             final Long scheduledStartTime = getScheduledStartTime(userMessageLog);
-            final int timeOut = legConfiguration.getReceptionAwareness().getRetryTimeout() * 60000;
+            final long timeOut = legConfiguration.getReceptionAwareness().getRetryTimeout() * 60000L;
             Date result = new Date(scheduledStartTime + timeOut);
             LOG.debug("Message expiration date is [{}]", result);
             return result;
@@ -305,7 +305,7 @@ public class UpdateRetryLoggingService {
     }
 
     public boolean isExpired(LegConfiguration legConfiguration, UserMessageLog userMessageLog) {
-        int delay = domibusPropertyProvider.getIntegerProperty(MESSAGE_EXPIRATION_DELAY);
+        long delay = domibusPropertyProvider.getLongProperty(MESSAGE_EXPIRATION_DELAY);
         Boolean isExpired = (getMessageExpirationDate(userMessageLog, legConfiguration).getTime() + delay) < System.currentTimeMillis();
         LOG.debug("Verify if message expired: [{}]", isExpired);
         return isExpired;
